6-size.c

+++ 0x00-hello_world/6-size.c
@@ -5,8 +5,8 @@
  */
 int main(void)
 {
-	printf("size ofa char: %zu byte(s)\n" , sizeof(char));
-	printf("size of an int: %zu byte(s)\n", sizeof(long int));
+	printf("size ofa char: %zu byte(s)\n", sizeof(char));
+	printf("size of an int: %zu byte(s)\n", sizeof(int));
 	printf("size of a long int: %zu byte(s)\n", sizeof(long int));
 	printf("size of a float: %zu byte(s)\n", sizeof(float));
 	return (0)
